+++ tests/lib.sh
@@ -2,8 +2,6 @@
 
 set -o pipefail
 
-BUILD_JAVA_LIBRARY="${BUILD_JAVA_LIBRARY:-false}"
-
 HARDWIRED_MOCK_TOP=
 # Uncomment this to be able to view the source tree after a test is run
 # HARDWIRED_MOCK_TOP=/tmp/td
@@ -11,19 +9,18 @@ HARDWIRED_MOCK_TOP=
 REAL_TOP="$(readlink -f "$(dirname "$0")"/../../..)"
 
 function make_mock_top {
-  mock=$(mktemp -t -d st.XXXXX)
+  mock=$(mktemp -t -d st.XXXXXX)
   echo "$mock"
 }
 
+MOCK_TOP_TO_CLEAN_UP=()
+WARMED_UP_MOCK_TOP_TO_CLEAN_UP=()
 if [[ -n "$HARDWIRED_MOCK_TOP" ]]; then
   MOCK_TOP="$HARDWIRED_MOCK_TOP"
-else
+elif [[ -z "${MOCK_TOP:-}" ]]; then
   MOCK_TOP=$(make_mock_top)
-  trap cleanup_mock_top EXIT
+  MOCK_TOP_TO_CLEAN_UP+=("$MOCK_TOP")
 fi
-
-WARMED_UP_MOCK_TOP=$(mktemp -t soong_integration_tests_warmup.XXXXXX.tar.gz)
-trap 'rm -f "$WARMED_UP_MOCK_TOP"' EXIT
 
 function warmup_mock_top {
   info "Warming up mock top ..."
@@ -32,19 +29,52 @@ function warmup_mock_top {
   mkdir -p "$MOCK_TOP"
   cd "$MOCK_TOP"
 
-  if [[ "$BUILD_JAVA_LIBRARY" == "true" ]]; then
-    create_mock_soong_for_java_lib
-    run_soong_for_java_lib
-  else
-    create_mock_soong
-    run_soong
-  fi
-  tar czf "$WARMED_UP_MOCK_TOP" *
+  create_mock_soong
+  run_soong
+
+  tar cf "$WARMED_UP_MOCK_TOP" --exclude out/.path_interposer_log *
+}
+
+function extend_mock_top {
+  local old_mock_top="$WARMED_UP_MOCK_TOP"
+  WARMED_UP_MOCK_TOP=$(mktemp -t soong_integration_tests_warmup.tar.XXXXXX)
+  WARMED_UP_MOCK_TOP_TO_CLEAN_UP+=("$WARMED_UP_MOCK_TOP")
+
+  info "Extending mock top with $1 ..."
+  info "New mock top archive: $WARMED_UP_MOCK_TOP"
+
+  cleanup_mock_top
+  mkdir -p "$MOCK_TOP"
+  cd "$MOCK_TOP"
+
+  tar xf "$old_mock_top"
+  $1
+  tar cf "$WARMED_UP_MOCK_TOP" --exclude out/.path_interposer_log *
 }
 
 function cleanup_mock_top {
   cd /
-  rm -fr "$MOCK_TOP"
+  if [[ -n "$MOCK_TOP" && -d "$MOCK_TOP" ]]; then
+    rm -fr "$MOCK_TOP"
+  fi
+}
+
+function cleanup_all {
+  cd /
+  if [[ -n "${MOCK_TOP_TO_CLEAN_UP:-}" ]]; then
+    for m in ${MOCK_TOP_TO_CLEAN_UP[@]}; do
+      if [[ -d "$m" ]]; then
+        rm -fr "$m"
+      fi
+    done
+  fi
+  if [[ -n "${WARMED_UP_MOCK_TOP_TO_CLEAN_UP:-}" ]]; then
+    for m in ${WARMED_UP_MOCK_TOP_TO_CLEAN_UP[@]}; do
+      if [[ -f "$m" ]]; then
+        rm -fr "$m"
+      fi
+    done
+  fi
 }
 
 function info {
@@ -102,92 +132,21 @@ function create_mock_soong {
   copy_directory build/soong
   copy_directory build/make
 
-  symlink_directory prebuilts/sdk
-  symlink_directory prebuilts/go
-  symlink_directory prebuilts/build-tools
-  symlink_directory prebuilts/clang/host
   symlink_directory external/compiler-rt
   symlink_directory external/go-cmp
   symlink_directory external/golang-protobuf
   symlink_directory external/licenseclassifier
-  symlink_directory external/starlark-go
-  symlink_directory external/python
-  symlink_directory external/sqlite
-  symlink_directory external/spdx-tools
-  symlink_directory libcore
-
-  # TODO: b/286872909 - Remove these when the blocking bug is completed
-  symlink_directory external/libavc
-  symlink_directory external/libaom
-  symlink_directory external/libvpx
-  symlink_directory frameworks/base/libs/androidfw
-  symlink_directory external/libhevc
-  symlink_directory external/libexif
-  symlink_directory external/libopus
-  symlink_directory external/libmpeg2
-  symlink_directory external/expat
-  symlink_directory external/flac
-  symlink_directory system/extras/toolchain-extras
-
-  touch "$MOCK_TOP/Android.bp"
-}
-
-function create_mock_soong_for_java_lib {
-  copy_directory build/blueprint
-  copy_directory build/soong
-  copy_directory build/make
-
-  symlink_directory prebuilts/sdk
-  symlink_directory prebuilts/go
-  symlink_directory prebuilts/build-tools
-  symlink_directory prebuilts/clang/host
-  symlink_directory external/compiler-rt
-  symlink_directory external/go-cmp
-  symlink_directory external/golang-protobuf
-  symlink_directory external/licenseclassifier
-  symlink_directory external/starlark-go
-  symlink_directory external/python
-  symlink_directory external/sqlite
-  symlink_directory external/spdx-tools
-  symlink_directory libcore
-
-  # TODO: b/286872909 - Remove these when the blocking bug is completed
-  symlink_directory external/libavc
-  symlink_directory external/libaom
-  symlink_directory external/libvpx
-  symlink_directory frameworks/base/libs/androidfw
-  symlink_directory external/libhevc
-  symlink_directory external/libexif
-  symlink_directory external/libopus
-  symlink_directory external/libmpeg2
-  symlink_directory external/expat
-  symlink_directory external/flac
-  symlink_directory system/extras/toolchain-extras
-
-  # Required for building java_library type modules.
-  symlink_directory prebuilts/jdk
-  symlink_directory prebuilts/r8
-  symlink_directory prebuilts/rust
-  symlink_directory prebuilts/gcc
-  symlink_directory external/turbine
-  symlink_directory external/jarjar
-  symlink_directory external/rust
-  symlink_directory external/gson
-  symlink_directory external/ow2-asm
+  symlink_directory external/pogreb
   symlink_directory external/protobuf
-  symlink_directory external/auto
-  symlink_directory external/icu
-  symlink_directory external/zlib
-  symlink_directory external/guava
-  symlink_directory external/jspecify
-  symlink_directory external/error_prone
-  symlink_directory external/jsr305
-  symlink_directory external/javapoet
-  symlink_directory external/escapevelocity
-  symlink_directory external/kotlinx.metadata
-  symlink_directory external/kotlinc
-  symlink_directory prebuilts/misc/common/asm/
-  symlink_directory system/logging/liblog
+  symlink_directory external/python
+  symlink_directory external/spdx-tools
+  symlink_directory external/sqlite
+  symlink_directory external/starlark-go
+  symlink_directory prebuilts/build-tools
+  symlink_directory prebuilts/siso
+  symlink_directory prebuilts/clang/host
+  symlink_directory prebuilts/go
+  symlink_directory prebuilts/sdk
 
   touch "$MOCK_TOP/Android.bp"
 }
@@ -201,117 +160,41 @@ function setup {
   info "Running test case \e[96;1m${FUNCNAME[1]}\e[0m"
   cd "$MOCK_TOP"
 
-  tar xzf "$WARMED_UP_MOCK_TOP" --warning=no-timestamp
-}
-
-# This file is generated by soong during make, here we have copy pasted the
-# exact same file, with VendorApiLevel added as an extra.
-# shellcheck disable=SC2120
-function write_soong_vars {
-  local soong_out=$MOCK_TOP/out/soong
-  local variables_file_name="$1"
-  local tmp_variables_file="$soong_out"/"$variables_file_name".tmp
-  local variables_file="$soong_out"/"$variables_file_name"
-  mkdir -p "$soong_out"
-      cat > "$tmp_variables_file" << EOF
-{
-    "BuildNumberFile": "build_number.txt",
-    "Platform_version_name": "S",
-    "Platform_sdk_version": 30,
-    "Platform_sdk_codename": "S",
-    "Platform_sdk_final": false,
-    "Platform_base_sdk_extension_version": 30,
-    "Platform_version_active_codenames": [
-        "S"
-    ],
-    "Platform_version_all_preview_codenames": [
-        "S"
-    ],
-    "DeviceName": "generic_arm64",
-    "DeviceProduct": "aosp_arm-eng",
-    "DeviceArch": "arm64",
-    "DeviceArchVariant": "armv8-a",
-    "DeviceCpuVariant": "generic",
-    "DeviceAbi": [
-        "arm64-v8a"
-    ],
-    "DeviceMaxPageSizeSupported": "4096",
-    "DeviceNoBionicPageSizeMacro": false,
-    "DeviceSecondaryArch": "arm",
-    "DeviceSecondaryArchVariant": "armv8-a",
-    "DeviceSecondaryCpuVariant": "generic",
-    "DeviceSecondaryAbi": [
-        "armeabi-v7a",
-        "armeabi"
-    ],
-    "HostArch": "x86_64",
-    "HostSecondaryArch": "x86",
-    "CrossHost": "windows",
-    "CrossHostArch": "x86",
-    "CrossHostSecondaryArch": "x86_64",
-    "AAPTCharacteristics": "nosdcard",
-    "AAPTConfig": [
-        "normal",
-        "large",
-        "xlarge",
-        "hdpi",
-        "xhdpi",
-        "xxhdpi"
-    ],
-    "AAPTPreferredConfig": "xhdpi",
-    "AAPTPrebuiltDPI": [
-        "xhdpi",
-        "xxhdpi"
-    ],
-    "Malloc_low_memory": false,
-    "Malloc_zero_contents": true,
-    "Malloc_pattern_fill_contents": false,
-    "Safestack": false,
-    "Build_from_text_stub": false,
-    "BootJars": [],
-    "ApexBootJars": [],
-    "PartitionVarsForSoongMigrationOnlyDoNotUse": {
-        "PartitionQualifiedVariables": null
-    },
-    "CompatibilityTestcases": null,
-    "VendorApiLevel": "202604"
-}
-EOF
-if cmp -s "$tmp_variables_file" "$variables_file"; then
-    rm "$tmp_variables_file"
-else
-    mv -f "$tmp_variables_file" "$variables_file"
-fi
+  tar xf "$WARMED_UP_MOCK_TOP"
 }
 
 # shellcheck disable=SC2120
 function run_soong {
-  USE_RBE=false TARGET_PRODUCT=aosp_arm TARGET_RELEASE=trunk_staging TARGET_BUILD_VARIANT=userdebug build/soong/soong_ui.bash --make-mode --skip-ninja --skip-config --soong-only --skip-soong-tests "$@"
-}
-
-# shellcheck disable=SC2120
-function run_soong_for_java_lib {
-  write_soong_vars soong.aosp_arm.variables
-  write_soong_vars soong.variables
-  USE_RBE=false TARGET_PRODUCT=aosp_arm TARGET_RELEASE=trunk_staging TARGET_BUILD_VARIANT=eng build/soong/soong_ui.bash --make-mode --skip-ninja --skip-config --soong-only --skip-soong-tests "$@"
+  OUT_DIR=out USE_RBE=false TARGET_PRODUCT=test_arm64 TARGET_RELEASE=trunk_staging TARGET_BUILD_VARIANT=eng build/soong/soong_ui.bash --make-mode --skip-ninja --skip-config --soong-only --skip-soong-tests "$@"
 }
 
 function run_ninja {
-  build/soong/soong_ui.bash --make-mode --skip-config --soong-only --skip-soong-tests "$@"
+  OUT_DIR=out USE_RBE=false TARGET_PRODUCT=test_arm64 TARGET_RELEASE=trunk_staging TARGET_BUILD_VARIANT=eng build/soong/soong_ui.bash --make-mode --skip-config --soong-only --skip-soong-tests "$@"
 }
 
 info "Starting Soong integration test suite $(basename "$0")"
 info "Mock top: $MOCK_TOP"
 
+trap cleanup_all EXIT
 
 export ALLOW_MISSING_DEPENDENCIES=true
 export ALLOW_BP_UNDER_SYMLINKS=true
-warmup_mock_top
+
+if [[ -z "${WARMED_UP_MOCK_TOP:-}" ]]; then
+  WARMED_UP_MOCK_TOP=$(mktemp -t soong_integration_tests_warmup.tar.gz.XXXXXX)
+  warmup_mock_top
+  WARMED_UP_MOCK_TOP_TO_CLEAN_UP+=("$WARMED_UP_MOCK_TOP")
+fi
 
 function scan_and_run_tests {
+  test_fns=("$@")
   # find all test_ functions
   # NB "declare -F" output is sorted, hence test order is deterministic
-  readarray -t test_fns < <(declare -F | sed -n -e 's/^declare -f \(test_.*\)$/\1/p')
+  if [[ ${#test_fns[*]} -eq 0 ]]; then
+    while IFS= read test_fn; do
+        test_fns+=($test_fn)
+    done < <(declare -F | sed -n -e 's/^declare -f \(test_.*\)$/\1/p')
+  fi
   info "Found ${#test_fns[*]} tests"
   if [[ ${#test_fns[*]} -eq 0 ]]; then
     fail "No tests found"
@@ -322,10 +205,79 @@ function scan_and_run_tests {
   done
 }
 
-function move_mock_top {
-  MOCK_TOP2=$(make_mock_top)
-  rm -rf $MOCK_TOP2
-  mv $MOCK_TOP $MOCK_TOP2
-  MOCK_TOP=$MOCK_TOP2
-  trap cleanup_mock_top EXIT
+function assert_files_equal {
+  if [ $# -ne 2 ]; then
+    echo "Usage: assert_files_equal file1 file2"
+    exit 1
+  fi
+
+  if ! cmp -s "$1" "$2"; then
+    echo "Files are different: $1 $2"
+    diff -u "$1" "$2"
+    exit 1
+  fi
+}
+
+function compare_incremental_files() {
+  local dir_full=$1; shift
+  local dir_incremental=$1; shift
+  count=0
+  for file_before in ${dir_full}/*.ninja*; do
+    basename=$(basename "$file_before")
+    file_after="${dir_incremental}/${basename}"
+    # The incremental file should be a superset of the full one.
+    if [[ "$basename" == "build.test_arm64.ninja" ]]; then
+      echo "Performing superset check for $basename..."
+      extra_lines=$(comm -23 <(sort "$file_before") <(sort "$file_after"))
+      if [[ -n "$extra_lines" ]]; then
+        # If there are extra lines, print an error and the differing lines, then exit.
+        echo "ERROR: $file_after is NOT a superset of $file_before."
+        echo "The following lines were in the 'before' file but not the 'after' file:"
+        echo "$extra_lines"
+        exit 1
+      fi
+    elif [[ "${basename}" == "build.test_arm64.ninja.globs_time" ]]; then
+      : # skip timestamp file
+    elif [[ "${basename}" == "build.test_arm64.ninja.glob_results" ]]; then
+      : # skip timestamp file
+    else
+      assert_files_equal $file_before $file_after
+    fi
+    ((count++)) || true
+  done
+  echo "Compared $count ninja files"
+}
+
+function compare_files_parity() {
+  local dir_before=$1; shift
+  local dir_after=$1; shift
+  count=0
+  for file_before in ${dir_before}/*.*; do
+    basename="$(basename "${file_before}")"
+    file_after="${dir_after}/${basename}"
+    if [[ "${basename}" == "build.test_arm64.ninja.globs_time" ]]; then
+      : # skip timestamp file
+    elif [[ "${basename}" == "build.test_arm64.ninja.glob_results" ]]; then
+      : # skip timestamp file
+    else
+    assert_files_equal $file_before $file_after
+    ((count++)) || true
+    fi
+  done
+  echo "Compared $count ninja files"
+}
+
+# Assumes an initial run of "run_soong SOONG_INCREMENTAL_ANALYSIS=true" and then
+# the tree is modified.
+function compare_incremental_and_full_analysis() {
+    run_soong SOONG_INCREMENTAL_ANALYSIS=true "$@"
+    mkdir incremental
+    cp -pr out/soong/build.test_arm64*.ninja* incremental
+
+    touch Android.bp
+    run_soong SOONG_INCREMENTAL_ANALYSIS=false "$@"
+    mkdir full
+    cp -pr out/soong/build.test_arm64*.ninja* full
+
+    compare_incremental_files full incremental
 }
